+++ packages/synthetix-main/contracts/interfaces/external/IRewardDistributor.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.0;
 interface IRewardDistributor {
     /// called by system
     function payout(
-        uint poolId,
+        uint128 poolId,
         address token,
         address to,
         uint amount
