+++ packages/synthetix-main/contracts/mocks/IAccountRBACMixinModuleMock.sol
@@ -2,9 +2,9 @@
 pragma solidity ^0.8.0;
 
 interface IAccountRBACMixinModuleMock {
-    function mockAccountRBACMixinDeposit(uint accountId, uint newDepositMock) external;
+    function mockAccountRBACMixinDeposit(uint128 accountId, uint newDepositMock) external;
 
-    function mockAccountRBACMixinMint(uint accountId, uint newMintMock) external;
+    function mockAccountRBACMixinMint(uint128 accountId, uint newMintMock) external;
 
     function mockAccountRBACMixinGetDepositMock() external view returns (uint);
